fix image-cri-shim config

+++ k3s/scripts/clean-k3s.sh
@@ -12,11 +12,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+set -x
+
 cd "$(dirname "$0")" >/dev/null 2>&1 || exit
 source common.sh
-systemctl stop k3s
-systemctl disable k3s
-systemctl daemon-reload
 
 rm -f /usr/bin/k3s-pre-start.sh
 rm -f /usr/bin/k3s-post-stop.sh
@@ -30,15 +29,46 @@ if grep -E "($sealos_b|$sealos_e)" /etc/security/limits.conf >/dev/null 2>&1; th
   sed -i "${slb},${sle}d" /etc/security/limits.conf
 fi
 
-rm -rf /usr/bin/{kubectl,crictl,ctr,k3s}
+
+SYSTEM_NAME=k3s
+
+BIN_DIR=/usr/local/bin
+
+${BIN_DIR}/k3s-killall.sh
+
+if command -v systemctl; then
+    systemctl disable ${SYSTEM_NAME}
+    systemctl reset-failed ${SYSTEM_NAME}
+    systemctl daemon-reload
+fi
+if command -v rc-update; then
+    rc-update delete ${SYSTEM_NAME} default
+fi
+
 rm -f /etc/systemd/system/k3s.service
-rm -rf /etc/systemd/system/k3s.service.d
+rm -f /etc/rancher/k3s/*.env
+
+remove_uninstall() {
+    rm -f ${BIN_DIR}/k3s-uninstall.sh
+}
+trap remove_uninstall EXIT
+
+if (ls /etc/systemd/system/k3s*.service || ls /etc/init.d/k3s*) >/dev/null 2>&1; then
+    set +x; echo 'Additional k3s services installed, skipping uninstall of k3s'; set -x
+    exit
+fi
+
+for cmd in kubectl crictl ctr; do
+    if [ -L ${BIN_DIR}/$cmd ]; then
+        rm -f ${BIN_DIR}/$cmd
+    fi
+done
+
 rm -rf /etc/rancher/k3s
 rm -rf /run/k3s
 rm -rf /run/flannel
 rm -rf /var/lib/rancher/k3s
 rm -rf /var/lib/kubelet
-
-bash killall.sh
+rm -f ${BIN_DIR}/k3s
 
 logger "clean kubelet success"
